feat: extract TaskStore into pluggable store abstraction layer
Mirrors the existing transport/ pattern (Transport ABC + FileTransport +
factory). Existing imports (`from clawteam.team.tasks import TaskStore`)
keep working via lazy re-export.

New layout:
  store/base.py   — BaseTaskStore ABC with get_stats default impl
  store/file.py   — FileTaskStore (moved from team/tasks.py, unchanged)
  store/__init__   — get_task_store() factory, reads CLAWTEAM_TASK_STORE
  config.py       — task_store field + env var mapping

17 new tests, 201 total passing.
